chore (search-api) Move controller classes to controller package
