+++ applications/search-api/src/main/java/no/dcat/portal/query/controller/HarvestQueryService.java
@@ -1,7 +1,8 @@
-package no.dcat.portal.query;
+package no.dcat.portal.query.controller;
 
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
+import no.dcat.portal.query.ElasticsearchService;
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.index.query.QueryBuilder;
