+++ applications/search-api/src/test/java/no/dcat/portal/query/controller/DatasetsQueryServiceDetailTest.java
@@ -1,5 +1,7 @@
-package no.dcat.portal.query;
+package no.dcat.portal.query.controller;
 
+import no.dcat.portal.query.ElasticsearchService;
+import no.dcat.portal.query.ServletRequest;
 import no.dcat.shared.Dataset;
 import no.fdk.test.testcategories.UnitTest;
 import no.fdk.webutils.exceptions.NotFoundException;
